+++ examples/path-aliases-v1/pages/about/+config.h.ts
@@ -5,7 +5,7 @@ export default {
   // https://vike.dev/meta
   meta: {
     Page: {
-      env: 'server-only'
+      env: { server: true }
     }
   }
 } satisfies Config