+++ vike/node/plugin/plugins/devConfig/determineOptimizeDeps.ts
@@ -5,7 +5,7 @@ import { findPageFiles } from '../../shared/findPageFiles.js'
 import { assert, getFilePathAbsolute, isNotNullish, isNpmPackageImport, unique } from '../../utils.js'
 import { getVikeConfig } from '../importUserCode/v1-design/getVikeConfig.js'
 import { ConfigVikeResolved } from '../../../../shared/ConfigVike.js'
-import { getConfigValueSourcesRelevant } from '../../shared/getConfigValueSourcesRelevant.js'
+import { getConfigValueSourcesNotOverriden } from '../../shared/getConfigValueSourcesNotOverriden.js'
 import { analyzeClientEntries } from '../buildConfig.js'
 import type { PageConfigBuildTime } from '../../../../shared/page-configs/PageConfig.js'
 import {
@@ -44,12 +44,11 @@ async function getPageDeps(config: ResolvedConfig, pageConfigs: PageConfigBuildT
   // V1 design
   {
     pageConfigs.forEach((pageConfig) => {
-      const configValueSourcesRelevant = getConfigValueSourcesRelevant(pageConfig)
-      configValueSourcesRelevant.forEach((configValueSource) => {
+      getConfigValueSourcesNotOverriden(pageConfig).forEach((configValueSource) => {
         if (!configValueSource.valueIsImportedAtRuntime) return
         const { definedAt, configEnv } = configValueSource
 
-        if (configEnv !== 'client-only' && configEnv !== 'server-and-client') return
+        if (!configEnv.client) return
 
         if (definedAt.filePathRelativeToUserRootDir !== null) {
           const { filePathAbsoluteFilesystem } = definedAt